rename : 파일명 통일화 remove : 사용하지 않는 이미지 삭제
