+++ src/components/party/HostPartyList.jsx
@@ -4,7 +4,7 @@ import { PARTIES_URL } from '../../shared/constants';
 import HostPartyItem from './HostPartyItem';
 import { styled } from 'styled-components';
 import Default from '../user/Default';
-import { ReactComponent as DefaultIcon } from '../../assets/mypage/default2.svg';
+import { ReactComponent as DefaultIcon } from '../../assets/mypage/no-party.svg';
 import Loading from '../common/Loading';
 import { ReactComponent as Backicon } from '../../assets/userprofile/back.svg';
 import { useNavigate } from 'react-router-dom';
