+++ src/components/party/PartyApply.jsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 
 // 이미지 import
 import { ReactComponent as ApproveBack } from '../../assets/common/close.svg';
-import { ReactComponent as Check } from '../../assets/party/approvecheck.svg';
+import { ReactComponent as Check } from '../../assets/party/approve-check.svg';
 import { ReactComponent as Warning } from '../../assets/common/information.svg';
 import queryString from 'query-string';
 import { postAPI } from '../../api/api';
